fix(connlib): log fragmented IP packets on debug (#10488)
When an application sends UDP packets that are larger than the MTU of
the underlying interface, the kernel fragments the packet at the IP
level. Firezone does not support fragmented IP packets because we need
to pack each IP packet into a UDP packet.

Right now, we don't check for fragmented IP packets which results in
packet parsing errors because the slice we are trying to parse the
packet from is not long enough.

To avoid spamming Sentry in these cases, we explicitly check for
fragmented IP packets and only log those on DEBUG.

Resolves: #10335