chore(portal): remove Jason in favor of JSON
Since Elixir 1.18, json encoding and decoding support is included in the standard library. This is build on OTP's native json support which is often faster than other implementations.

It mostly has the same API as the popular Jason library, differing mainly in the format of the error responses returned when decoding fails.

To minimize dependence on external libraries, we remove the Jason lib in favor of this external dependency.