chore(website): Use `bundler` for module resolution
CustomFlowbiteTheme was restructured and we needed to add `/types`
suffix. This triggered the following:

```
Cannot find module 'flowbite-react/types' or its corresponding type declarations.
Consider updating to 'node16', 'nodenext', or 'bundler'.
```

According to TypeScript 5.0 docs:

> "If you are using a modern bundler like Vite, esbuild, swc, Webpack,
Parcel, and others that implement a hybrid lookup strategy, the new
bundler option should be a good fit for you."

See:
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-0.html

Note: we use Next.js, which is not a bundler itself, but it uses Webpack.