fix(portal): trigger email for Okta 4xx errors
When Okta returned a 4xx status code from the API, we had updated error handler to grab the errors from body or headers and returning these.

However, the caller was expecting an explicit empty string for 401 and 403 errors in order to trigger the email send behavior.

Since that wasn't being matched, we were logging the error internally only, and continuing to retry the sync indefinitely without sending the user an email.